+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 - Handle `MemoryError`
 - Fix `execute_battle_with_retry` return value
 - Refactor API exception classes
-- Upgrade resources to `v0488`
+- Upgrade resources to `v0490`
 - Add mission APIs
 - Farm Hall of Fame reward on Saturday
 - Disable `backtrace`